Fixed bug with API that going from colorloop to none would reset lamp
Due to the inner workings of the API, if the bridge receives a
PUT-request while in a colorloop (and perharps future effects), the lamp
will handle this change last. When doing so, it will reset the lamp to
the last color before the colorloop, and not the new one being set. This
handles this by always setting effects to none in its' own request.